requirements.txt
pillow
selenium
python_dotenv
webdriver_manager
requests
beautifulsoup4